divided Dialog Device Dept
