+++ src/layouts/DepartmentTable/Dialog/index.ts
@@ -1,0 +1,7 @@
+export {default as DialogCreate } from './DialogCreate'
+export {default as DialogDelete} from './DialogDelete'
+export {default as DialogDeviceUsageHours} from './DialogDeviceUsageHours'
+export {default as DialogEdit} from './DialogEdit'
+export {default as DialogHistoryDevices} from './DialogHistoryDevices'
+export {default as DialogImportDeviceInfo} from './DialogImportDeviceInfo'
+export {default as DialogMaintenanceDevice} from './DialogMaintenanceDevice'
