+++ src/pages/appSlice.tsx
@@ -11,8 +11,8 @@ interface ISnackbarMessage {
 export const defaultSnackbarMessage: ISnackbarMessage = {
 	isOpen: false,
 	message: '',
-	color: "black",
-	backgroundColor: 'white'
+	color: 'black',
+	backgroundColor: 'white',
 };
 
 interface ISidebarItem {
@@ -110,7 +110,12 @@ export const defaultSidebarItems: ISidebarItem[] = [
 	},
 	{
 		isOpen: false,
-		name: 'Điều chuyển thiết bị',
+		name: 'Phiếu đăng ký',
+		icon: '',
+	},
+	{
+		isOpen: false,
+		name: 'Chuyển đổi thiết bị',
 		icon: '',
 	},
 ];
@@ -158,7 +163,7 @@ export const appSlice = createSlice({
 				...state,
 				snackbarState: {
 					isOpen: action.payload ? true : false,
-					message: action.payload ? action.payload : '',	
+					message: action.payload ? action.payload : '',
 				},
 			};
 		},
@@ -167,20 +172,17 @@ export const appSlice = createSlice({
 				...state,
 				snackbarState: {
 					isOpen: action.payload ? true : false,
-					message: action.payload ? action.payload.message : '',	
-					color:  action.payload.hasOwnProperty('color') ?  action.payload.color : state.snackbarState.color,
-					backgroundColor: action.payload.hasOwnProperty('backgroundColor') ?  action.payload.backgroundColor : state.snackbarState.backgroundColor
+					message: action.payload ? action.payload.message : '',
+					color: action.payload.hasOwnProperty('color') ? action.payload.color : state.snackbarState.color,
+					backgroundColor: action.payload.hasOwnProperty('backgroundColor')
+						? action.payload.backgroundColor
+						: state.snackbarState.backgroundColor,
 				},
 			};
 		},
 	},
 });
 
-export const { 
-	setIsOpenDrawer,
-	setSnackbarMessage, 
-	setSidebarItems,
-	setSnackbar,
- } = appSlice.actions;
+export const { setIsOpenDrawer, setSnackbarMessage, setSidebarItems, setSnackbar } = appSlice.actions;
 
 export default appSlice.reducer;
