+++ src/types/instrumentType.ts
@@ -50,3 +50,11 @@ export const dummyDeviceSpecData: IDeviceSpecType = {
     "SpecsName": "",
     "SpecsValue": ""
 }
+
+export interface ILiquidateDeptInstrument {
+    "InstrumentDeptId": String,
+    "DeviceId": String,
+    "DeviceName": String,
+    "Unit": String,
+    "QuantityTotal": Number
+}
